多线程部分基本完成

+++ database.py
@@ -1,30 +1,35 @@
-from typing import List, Optional, Sequence, Tuple
+from typing import Optional, Sequence, Tuple
 from config import *
 import sqlite3
 from contextlib import closing
+from unit import videoBaseInfoTuple, videoBindInfoTuple
 
 
-# TODO:last_watch_time
 def initDB():
     if os.path.exists(DB_PATH):
         return
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
-            cursor.execute("CREATE TABLE IF NOT EXISTS Video (hash TEXT PRIMARY KEY, fileName TEXT, filePath TEXT, fileSize TEXT, videoDuration TEXT)")
+            cursor.execute("CREATE TABLE IF NOT EXISTS Video (hash TEXT PRIMARY KEY, fileName TEXT, filePath TEXT, fileSize TEXT, videoDuration TEXT, lastWatchTime INTEGER DEFAULT -1)")
             cursor.execute("CREATE TABLE IF NOT EXISTS Binding (hash TEXT PRIMARY KEY, animeId INTEGER, episodeId INTEGER, animeTitle TEXT, episodeTitle TEXT, type TEXT, typeDescription TEXT)")
             connection.commit()
 
-def clearDB():
+
+def clearDB(part: str = 'all') -> None:
+    '''part: 'all' or 'video' or 'binding'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
-            cursor.execute("DELETE FROM Video")
-            cursor.execute("DELETE FROM Binding")
+            if part == 'all' or part == 'video':
+                cursor.execute("DELETE FROM Video")
+            if part == 'all' or part == 'binding':
+                cursor.execute("DELETE FROM Binding")
             connection.commit()
+
 
 def addVideoIntoDB(hash: str, fileName: str, filePath: str, fileSize: str, videoDuration: str) -> None:
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
-            cursor.execute("INSERT OR IGNORE INTO Video VALUES (?, ?, ?, ?, ?)", (hash, fileName, filePath, fileSize, videoDuration))
+            cursor.execute("INSERT OR IGNORE INTO Video(hash, fileName, filePath, fileSize, videoDuration) VALUES (?, ?, ?, ?, ?)", (hash, fileName, filePath, fileSize, videoDuration))
             connection.commit()
 
 
@@ -33,57 +38,66 @@ def addVideosIntoDB(videos: Sequence[Tuple[str, str, str, str, str]]) -> None:
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             for video in videos:
-                cursor.execute("INSERT OR IGNORE INTO Video VALUES (?, ?, ?, ?, ?)", video)
+                cursor.execute("INSERT OR IGNORE INTO Video(hash, fileName, filePath, fileSize, videoDuration) VALUES (?, ?, ?, ?, ?)", video)
             connection.commit()
 
-def addBindingIntoDB(hash: str, animeId: int, episodeId: int, animeTitle: str, episodeTitle: str, type: str, typeDescription: str) -> None:
+
+def addBindingIntoDB(hash: str, _videoBindInfoTuple: videoBindInfoTuple) -> None:
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
-            cursor.execute("INSERT OR IGNORE INTO Binding VALUES (?, ?, ?, ?, ?, ?, ?)", (hash, animeId, episodeId, animeTitle, episodeTitle, type, typeDescription))
+            cursor.execute("INSERT OR IGNORE INTO Binding VALUES (?, ?, ?, ?, ?, ?, ?)", (hash, *_videoBindInfoTuple[:6]))
             connection.commit()
 
-def addBindingsIntoDB(bindings: Sequence[Tuple[str, int, int, str, str, str, str]]) -> None:
-    '''eachTuple: [0] hash, [1] animeId, [2] episodeId, [3] animeTitle, [4] episodeTitle, [5] type, [6] typeDescription'''
+
+def addBindingsIntoDB(bindings: Sequence[Tuple[str, videoBindInfoTuple]]) -> None:
+    '''eachTuple: [0] hash, [1] videoBindInfoTuple'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             for binding in bindings:
-                cursor.execute("INSERT OR IGNORE INTO Binding VALUES (?, ?, ?, ?, ?, ?, ?)", binding)
+                cursor.execute("INSERT OR IGNORE INTO Binding VALUES (?, ?, ?, ?, ?, ?, ?)", (binding[0], *(binding[1][:6])))
             connection.commit()
 
-def getVideoFromDB(hash: str) -> Optional[Tuple[str, str, str, str, str]]:
-    '''[0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
+
+def getVideoFromDB(hash: str) -> Optional[videoBaseInfoTuple]:
+    '''videoBindInfoTuple: [0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Video WHERE hash=?", (hash,))
-            return cursor.fetchone()
+            return videoBaseInfoTuple._make(cursor.fetchone()[:5])
 
-def getAllVideos() -> List[Tuple[str, str, str, str, str]]:
-    '''eachTuple: [0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
+
+def getAllVideos() -> Tuple[videoBaseInfoTuple, ...]:
+    '''videoBindInfoTuple: [0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Video")
-            return cursor.fetchall()
+            return tuple(videoBaseInfoTuple._make(eachTuple[:5]) for eachTuple in cursor.fetchall())
+
 
 def checkIfVideoBinded(hash: str) -> bool:
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Binding WHERE hash=?", (hash,))
-            return bool(cursor.fetchone())
+            return cursor.fetchone() is None
 
-def getAllBindedVideos():
+# TODO: Logic mistake
+def getAllBindedVideos() -> Tuple[Tuple[str, videoBindInfoTuple], ...]:
+    '''[0] hash, [1] videoBindInfoTuple'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Binding")
-            return cursor.fetchall()
+            return tuple((eachTuple[0], videoBindInfoTuple(*(eachTuple[1:]))) for eachTuple in cursor.fetchall())
 
-def getAllUnBindedVideos() -> List[Tuple[str, str, str, str, str]]:
-    '''eachTuple: [0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
+
+def getAllUnBindedVideos() -> Tuple[videoBaseInfoTuple, ...]:
+    '''videoBaseInfoTuple: [0] hash, [1] fileName, [2] filePath, [3] fileSize, [4] videoDuration'''
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Video WHERE hash NOT IN (SELECT hash FROM Binding)")
-            return cursor.fetchall()
+            return tuple(videoBaseInfoTuple._make(eachTuple[:5]) for eachTuple in cursor.fetchall())
 
-def getBindingFromDB(hash: str):
+
+def getBindingFromDB(hash: str) -> Optional[videoBindInfoTuple]:
     with closing(sqlite3.connect(DB_PATH)) as connection:
         with closing(connection.cursor()) as cursor:
             cursor.execute("SELECT * FROM Binding WHERE hash=?", (hash,))
